Move GradleDistributionInfo to public package
