+++ org.eclipse.buildship.core/src/main/java/org/eclipse/buildship/core/projectimport/ProjectImportConfiguration.java
@@ -15,11 +15,11 @@ import java.io.File;
 import java.util.List;
 
 import org.eclipse.buildship.core.CorePlugin;
+import org.eclipse.buildship.core.GradleDistributionInfo;
 import org.eclipse.buildship.core.configuration.BuildConfiguration;
 import org.eclipse.buildship.core.util.binding.Property;
 import org.eclipse.buildship.core.util.binding.Validator;
 import org.eclipse.buildship.core.util.binding.Validators;
-import org.eclipse.buildship.core.util.gradle.GradleDistributionInfo;
 
 /**
  * Serves as the data model of the project import wizard.
