+++ org.eclipse.buildship.ui/src/main/java/org/eclipse/buildship/ui/preferences/GradleWorkbenchPreferencePage.java
@@ -19,11 +19,11 @@ import org.eclipse.ui.IWorkbenchPreferencePage;
 
 import org.eclipse.buildship.core.CorePlugin;
 import org.eclipse.buildship.core.GradleDistribution;
+import org.eclipse.buildship.core.GradleDistributionInfo;
 import org.eclipse.buildship.core.configuration.WorkspaceConfiguration;
 import org.eclipse.buildship.core.i18n.CoreMessages;
 import org.eclipse.buildship.core.util.binding.Validator;
 import org.eclipse.buildship.core.util.binding.Validators;
-import org.eclipse.buildship.core.util.gradle.GradleDistributionInfo;
 import org.eclipse.buildship.ui.util.font.FontUtils;
 import org.eclipse.buildship.ui.util.widget.GradleProjectSettingsComposite;
 import org.eclipse.buildship.ui.util.widget.GradleUserHomeGroup;
